Add files via upload
I have worked the logic out for the modify(), remove(), and insert() functions.
I have noticed that some additional functions that are called with the above, have been commented out, or do not exist. Should I work on creating the framework for these functions so that you can tweak them to working condition?